refactor(provider): ♻️ Rename fetch token helper file
