+++ src/index.ts
@@ -1,4 +1,4 @@
 export type { SapAiCoreModelId, SapAiCoreProvider, SapAiCoreProviderSettings } from './sap-ai-core-provider';
 export { createSapAiCore, sapAiCore } from './sap-ai-core-provider';
-export type { TokenServiceConfig } from './fetch-with-token';
-export { createFetchWithToken } from './fetch-with-token';
+export type { TokenServiceConfig } from './fetch-with-token-provider';
+export { createFetchWithToken } from './fetch-with-token-provider';
